stanczyk-server/requirements.txt
grpcio==1.34.0
grpcio-tools==1.34.0
numpy==1.19.5
opencv-python==4.5.1.48
protobuf==3.14.0
psutil==5.8.0
pytz==2020.5
six==1.15.0
tzlocal==2.1
